Improvements on aggregate signature proofs:
 * signatures are weightes by slots
 * pbft-proof as composite of prepare and commit proof

+++ primitives/block-albatross/src/view_change.rs
@@ -3,10 +3,10 @@ use beserial::{Deserialize, Serialize};
 use hash::SerializeContent;
 use super::signed;
 
-#[derive(Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Serialize, Deserialize, SerializeContent)]
+#[derive(Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Serialize, Deserialize, SerializeContent, Hash)]
 pub struct ViewChange {
-    pub block_number: u32,
-    pub new_view_number: u16,
+    pub block_number: u64,
+    pub new_view_number: u32,
 }
 
 impl signed::Message for ViewChange {
